+++ lib/src/features/purchase_orders/presentation/widgets/materials_view_bar_widget.dart
@@ -1,10 +1,12 @@
 import 'package:bflow_client/src/core/constants/colors.dart';
 import 'package:bflow_client/src/core/extensions/build_context_extensions.dart';
 import 'package:bflow_client/src/core/widgets/action_button_widget.dart';
+import 'package:bflow_client/src/core/widgets/confirmation_widget.dart';
 import 'package:bflow_client/src/features/jobs/presentation/bloc/job_bloc.dart';
 import 'package:bflow_client/src/features/purchase_orders/presentation/bloc/items_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:go_router/go_router.dart';
 
 import 'write_material_widget.dart';
 
@@ -42,8 +44,20 @@ class _MaterialsViewBarWidgetState extends State<MaterialsViewBarWidget> {
         Row(
           children: [
             ActionButtonWidget(
-              onPressed: () =>
-                  context.read<ItemsBloc>().add(DeleteItemsEvent()),
+              onPressed: () {
+                context.showCustomModal(
+                  ConfirmationWidget(
+                    title: "Delete materials",
+                    description:
+                        "Are you sure you want to delete the selected material(s)?",
+                    onConfirm: () {
+                      context.read<ItemsBloc>().add(DeleteItemsEvent());
+                      context.pop();
+                    },
+                    confirmText: "Delete",
+                  ),
+                );
+              },
               type: ButtonType.textButton,
               title: "Delete",
               icon: Icons.delete_outline,
